% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{analyze_results}
\alias{analyze_results}
\title{Summarizes classification results into list of 2 dataframes containing detailed
(per each position) and preliminary output (binary classification of reads for
modified - containing non-adenosine residue(s) and unmodified - with adenosines
exclusively).}
\usage{
analyze_results(coordinate_df, predicted_list)
}
\arguments{
\item{coordinate_df}{a data frame object produced by create_coordinate_df
function.}

\item{predicted_list}{a list object produced by predict_classes function.}
}
\value{
a list of 2 dataframes with prediction results. First dataframe contains
detailed info on type and estimated positions of non-adenosine residues detected.
The second dataframe contains initial indications, whether the given read was
recognized as modified (containing non-adenosine residue) or not.
}
\description{
Summarizes classification results into list of 2 dataframes containing detailed
(per each position) and preliminary output (binary classification of reads for
modified - containing non-adenosine residue(s) and unmodified - with adenosines
exclusively).
}
\examples{
\dontrun{

analyze_results(coordinate_df = '/path/to/coordinate_df',
                predicted_list = '/path/to/predicted_list')

}

}
