% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{assert_condition}
\alias{assert_condition}
\title{Assert condition is TRUE, stop with message if FALSE}
\usage{
assert_condition(cond, msg = "Assertion failed")
}
\arguments{
\item{cond}{Logical condition to check}

\item{msg}{Error message if condition is FALSE}
}
\value{
Invisible TRUE if assertion passes
}
\description{
Base R replacement for assertthat::assert_that
}
\keyword{internal}
