% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{calculate_fisher}
\alias{calculate_fisher}
\title{Performs Fisher's exact test for each transcript in ninetails
output data. Then, it performs the Benjamini-Hochberg procedure
(BH step-up procedure) to control the FDR.}
\usage{
calculate_fisher(
  ninetails_data,
  transcript_id_column = "contig",
  min_reads = 0,
  min_nonA_reads = 0,
  grouping_factor = "sample_name",
  condition1 = NA,
  condition2 = NA,
  alpha = 0.05,
  base = "C",
  ...
)
}
\arguments{
\item{ninetails_data}{dataframe - the output of ninetails::merge_nonA_tables
function (merged tabular output containing read classification &
non-A position data).}

\item{transcript_id_column}{[character string] column with transcript id data
(default: "contig", as inherited from nanopolish; can be changed by the user)}

\item{min_reads}{[numeric] minimum number of reads representing given
transcript to include it in the analysis. This parameter is set by default
to 0. Please keep in mind that taking into account many transcripts with low
coverage increases the risk of reject true null hypothesis
(Benjamini-Hochberg procedure).}

\item{min_nonA_reads}{[numeric] minimum number of reads containing
nonadenosine residues (summary for C, G, U alltogether) per given
transcript to include it in the analysis. This parameter prevents from
considering too many observations as nonsignificant in the further pvalue
adjustation. In general, the non-A containing reads are a small fraction
of the total pool of reads. As a rule of thumb, additional filtering
criteria can provide more valuable information regarding the samples
(prevent from rejecting true null hypothesis).
This is set by default to 0.}

\item{grouping_factor}{[character string] grouping variable
(e.g. "sample_name" - default)}

\item{condition1}{[character string] first level of `grouping_factor`
to use for comparison}

\item{condition2}{[character string] second level of `grouping_factor`
to use for comparison}

\item{alpha}{[numeric] an alpha value to consider a hit significant.
Default: 0.05.}

\item{base}{[character string] letter representing particular non-A nucleotide,
for which the statistics are meant to be computed. Currently function accepts
C, G, U arguments. The "C" value is set by default.}

\item{...}{additional parameters to pass to nonA_fisher (under development)}
}
\value{
a summary table with pvalues, padj and significance levels
for each transcript  (tibble)
}
\description{
This is a wrapper for fisher.test function from stats package and p.adjust
functions with additional features to facilitate data wrangling.
}
\details{
The function was inspired by the Nanotail package written & maintained by
Pawel Krawczyk (smaegol): https://github.com/LRB-IIMCB/nanotail/blob/dev/R/polya_stats.R

Many thanks to the developer of original source code.
}
\examples{
\dontrun{
test <- ninetails::calculate_fisher(ninetails_data=merged_nonA_tables,
                                    transcript_id_column = "contig",
                                    min_reads=100,
                                    min_nonA_reads=10,
                                    grouping_factor = "sample_name",
                                    condition1="WT",
                                    condition2="KO",
                                    alpha=0.05,
                                    base="C")
}
}
