% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfindr_compatibility.R
\name{check_polya_length_filetype}
\alias{check_polya_length_filetype}
\title{Check and convert poly(A) length file format}
\usage{
check_polya_length_filetype(input)
}
\arguments{
\item{input}{Path to file or data frame containing poly-A length data:
either produced by nanopolish or tailfindr. In case of tailfindr, only DRS
data are accepted.}
}
\value{
Standardized data frame in nanopolish-like format compatible with
further ninetails processing steps
}
\description{
This function checks whether the input poly(A) length files store data from
correct sequencing platform (Oxford Nanopore DRS sequencing) - if other data
types are loaded, an error occurs.
}
\details{
Please keep in mind:
This function is an element of legacy pipeline which would probably
be retired if fast5 format would be deprecated.
}
\examples{
\dontrun{

# run on nanopolish output
test <- ninetails::check_polya_length_filetype(
input = system.file('extdata',
                    'test_data',
                    'nanopolish_output.tsv',
                    package = 'ninetails'))

# run on tailfindr output
test <- ninetails::check_polya_length_filetype(
input = system.file('extdata',
                    'test_data',
                    'tailfindr_output.csv',
                    package = 'ninetails'))

}
}
