% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_pipeline_wrapper.R
\name{check_tails}
\alias{check_tails}
\title{Wrapper function for ninetails package.}
\usage{
check_tails(
  nanopolish,
  sequencing_summary,
  workspace,
  num_cores,
  basecall_group,
  pass_only,
  save_dir
)
}
\arguments{
\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the Fast5 file hierarchy from which the data should be extracted.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}

\item{save_dir}{character string. Full path of the directory where the output
files containing the tail composition information should be stored.}
}
\value{
A list containing tail information organized by the read ID
is returned. Always assign this returned list to a variable, otherwise
the long list will be printed to the console, which may crash your R session.
}
\description{
This function allows to perform all of the steps required to discover
nonadenosine nucleotides within the given dataset using ninetails.
Please keep in mind, that during computations the function creates large
segmentation data. Therefore it may be wise to split the nanopolish table
beforehand and then run this function on nanopolish table chunks.
}
\details{
The output of this function is a list of 3 dataframes, containing: a) binary
classification of reads satisfying filtering criteria*, b) detailed positional
info regarding all potential nonadenosine residues detected, c) info
regarding discarded reads (the reason why the following read has been omitted
from the analysis).

The filtering criteria applied by ninetails are as follows: move of value =1
present, qc_tag = "PASS" or "PASS" & "SUFFCLIP", length estimated
by nanopolish >=10 nt).
}
\examples{
\dontrun{

check_tails(nanopolish = '/path/to/file',
            sequencing_summary = '/path/to/file',
            workspace = '/path/to/guppy/workspace',
            num_cores = 10,
            basecalled_group = 'Basecall_1D_000',
            pass_only=TRUE,
            save_dir = '/directory/where/output/shall/be/stored')

}
}
