% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{correct_class_data}
\alias{correct_class_data}
\title{Corrects the classification of reads contained in class_data table}
\usage{
correct_class_data(residue_data_edited, class_data)
}
\arguments{
\item{residue_data_edited}{[dataframe] A dataframe or tibble containing
corrected nonA residue predictions produced by \code{\link{correct_residue_data}}
function.}

\item{class_data}{[dataframe] A dataframe or tibble containing read_classes
predictions made by ninetails pipeline}
}
\value{
A tibble with corrected class_data (potential artifacts are reclassified).
In this dataframe, in comparison to the original (raw) class_data, two additional
columns are present:\itemize{
\item corr_class - result of corrected classification
\item corr_comments - comment adjusted accordingly
}

For more details check documentation of \code{\link{create_outputs}} function
and \code{\link{reclassify_ninetails_data}} function.
}
\description{
Introduces additional columns to the class_data ('corr_comments' and
'corr_class'), with reclassification based on positional data
(tail lengths and distribution of non-A residuals).
}
\details{
Based on these columns, the classification can be adjusted, which minimizes
the risk of including nanopolish artifacts.

---CAUTION---

Reads that contain only non-A nucleotides that are likely to be nanopolish
artifacts are reclassified in the output table as "unmodified" ("corr_class"
column), and their comment is changed from "YAY" to "MPU" ("corr_comments" column).
The latter is due to maintain compatibility with tag system used in plotting
functions.

It is recommended that the output of the \code{\link{reclassify_ninetails_data}}
function be used in further analyses, which is optimized for this purpose.

In order to plot the output of this function directly, one should rename
'corr_comments' and 'corr_class' columns to 'comments' and 'class'.
}
\examples{
\dontrun{
# the output of function \code{\link{check_tails}} or \code{\link{create_outputs}}
# is required (below denoted as "results") to run the following example
# for details see the documentation for the \code{\link{check_tails}}
# and \code{\link{create_outputs}}

class_data_corrected <- ninetails::correct_class_data(residue_data_edited = residue_data_edited,
                                                      class_data=results[[1]])
}

}
