% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{correct_labels}
\alias{correct_labels}
\title{Correcting class recognition in ninetails data}
\usage{
correct_labels(df)
}
\arguments{
\item{df}{data frame with ninetails read classification results (class_data)
or the output of \code{\link{merge_nonA_tables}} function.}
}
\value{
corrected dataframe (with changed class labels)
}
\description{
This helper fnction solves issues with backcompatibility of ninetails'
 read class naming convention.
}
\details{
Previously, the read classes were named as "modified", "unmodified" and
 "unclassified" (pre-release versions). However, it has been changed
 to more precise terms: "decorated", "blank" and "unclassified", since the
 presence of non-A within poly(A) tail does not necessary have to be
 the result of post-transcriptional, cytoplasmic modification. It might be
 caused by the semi-templated tails or polymerase slippage during RNA
 synthesis/canonical tailing or caused by other potential yet undefined
 mechanisms.
}
\examples{
\dontrun{
class_data <- ninetails::correct_labels(class_data)
}

}
