% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{count_overlaps}
\alias{count_overlaps}
\title{Counts overlapping segments created per given nanopore signal.}
\usage{
count_overlaps(signal, segment, overlap)
}
\arguments{
\item{signal}{numeric vector corresponding to the given ONT signal.}

\item{segment}{numeric [1]. Length of the chunk(s) to be created.}

\item{overlap}{numeric [1]. Length of the overlap between the chunks.}
}
\value{
number of overlapping chunks per given tail.
}
\description{
Counts overlapping segments created per given nanopore signal.
}
\examples{
\dontrun{

count_overlaps(signal = signal_vector,
               segment = 100
               overlap = 10)
}


}
