% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{count_residues}
\alias{count_residues}
\title{Counts non-A residues found in nonadenosine_residues dataframe produced
by ninetails pipeline.}
\usage{
count_residues(residue_data, grouping_factor = NA)
}
\arguments{
\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline}

\item{grouping_factor}{grouping variable (e.g. "sample_name")}
}
\value{
A tibble with counts for each non-A residue
}
\description{
Process the information returned by ninetails found in the nonadenosine_residues
dataframe in the prediction column. It is important to highlight that this function
returns counts as summary of hits, it does not provide a summary per read.
}
\examples{
\dontrun{

residue_counted <- ninetails::count_residues(residue_data=out[[2]],
                                             grouping_factor=NA)
}

}
