% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{create_coordinate_dataframe}
\alias{create_coordinate_dataframe}
\title{Creates a dataframe with segmentation and tail length info required for
non-adenosine position estimation.}
\usage{
create_coordinate_dataframe(tail_feature_list, num_cores)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
a dataframe with 3 columns: read ID (readname), total number of segments
per given tail (total_chunk) and poly(A) tail length (tail_length) estimated by
nanopolish.
}
\description{
Creates a dataframe with segmentation and tail length info required for
non-adenosine position estimation.
}
\examples{
\dontrun{

create_coordinate_dataframe(tail_feature_list = '/path/to/tail_feature_list',
                            num_cores = 10)

}


}
