% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{create_gasf}
\alias{create_gasf}
\title{Converting ONT signal to gramian angular summation field.}
\usage{
create_gasf(tail_chunk)
}
\arguments{
\item{tail_chunk}{character string. Name of the given signal chunk within the
analyzed dataset.}
}
\value{
an array (100,100,1) with values representing ONT signal.
}
\description{
Converting ONT signal to gramian angular summation field.
}
\examples{
\dontrun{

create_gasf(tail_chunk = "1234-jhgfdr54-io643gju-ouy4378989")

}

}
