% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gasf_list.R
\name{create_gasf_list}
\alias{create_gasf_list}
\title{Creates list of gramian angular summation matrices produced based on
list of splitted tails (tail chunks).}
\usage{
create_gasf_list(tail_chunk_list, num_cores)
}
\arguments{
\item{tail_chunk_list}{character string. Full path of the list object produced
by create_chunk_list function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
A list of gasf matrices organized by the read ID_index
is returned. Always assign this returned list to a variable, otherwise
the long list will be printed to the console, which may crash your R session.
}
\description{
Creates list of gramian angular summation matrices produced based on
list of splitted tails (tail chunks).
}
\examples{
\dontrun{

create_gasf_list(feature_list = '/path/to/chunk_list', num_cores = 10)

}


}
