% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{create_outputs}
\alias{create_outputs}
\title{Creates the list object containing tabular outputs of ninetails pipeline.}
\usage{
create_outputs(
  tail_feature_list,
  tail_chunk_list,
  nanopolish,
  predicted_list,
  num_cores,
  pass_only = TRUE
)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{tail_chunk_list}{list object produced by create_tail_chunk_list
function.}

\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{predicted_list}{a list object produced by predict_classes function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}
}
\value{
This function returns a list object containing two fataframes:
"read_classes" and "nonadenosine_residues" with the final output.
First dataframe contains initial indications, whether the given read was
classified or omitted (with reason) and if classified, whether read was
recognized as modified (containing non-adenosine residue) or not.
The second dataframe contains detailed info on type and estimated positions
of non-adenosine residues detected.
}
\description{
Creates the list object containing tabular outputs of ninetails pipeline.
}
\examples{
\dontrun{

create_outputs(tail_feature_list = tail_feature_list,
               tail_chunk_list = tail_chunk_list,
               nanopolish = '/path/to/nanopolish_output.tsv',
               predicted_list = predicted_list,
               num_cores = 2,
               pass_only=TRUE)
}


}
