% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{create_tail_chunk_list}
\alias{create_tail_chunk_list}
\title{Extracts fragments of polyA tails of ONT RNA reads potentially containing
non-A nucleotides along their coordinates & appends the data to the nested
list organized by read IDs.}
\usage{
create_tail_chunk_list(tail_feature_list, num_cores)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
a list object (nested) containing the segmented tail data (chunks,
coordinates) organized by the read IDs.
}
\description{
Extracts fragments of polyA tails of ONT RNA reads potentially containing
non-A nucleotides along their coordinates & appends the data to the nested
list organized by read IDs.
}
\examples{
\dontrun{

create_tail_chunk_list(tail_feature_list = tail_feature_list,
                       num_cores = 2)

}
}
