% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tail_chunk_list_moved.R
\name{create_tail_chunk_list_moved}
\alias{create_tail_chunk_list_moved}
\title{Creates list of overlapping tail fragments extracted from
nanopolish output and fast5 files. Only fragments containing move==1
are included.}
\usage{
create_tail_chunk_list_moved(tail_feature_list, num_cores)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
A list of tail range chunks organized by the read ID
is returned. Only tails containing at least one move value equal to 1
are included. Always assign this returned list to a variable, otherwise
the long list will be printed to the console, which may crash your R session.
}
\description{
Creates list of overlapping tail fragments extracted from
nanopolish output and fast5 files. Only fragments containing move==1
are included.
}
\examples{
\dontrun{

create_tail_chunk_list_moved(tail_feature_list = '/path/to/tail_feature_list', num_cores = 10)

}


}
