% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions_dorado_cDNA.R
\name{extract_data_from_bam}
\alias{extract_data_from_bam}
\title{Extract data from BAM file for cDNA analysis}
\usage{
extract_data_from_bam(
  bam_file,
  summary_file,
  seq_only = TRUE,
  cli_log = message
)
}
\arguments{
\item{bam_file}{Character string. Path to BAM file containing basecalled sequences.}

\item{summary_file}{Character string. Path to corresponding Dorado summary file
containing read IDs to extract.}

\item{seq_only}{logical [TRUE]. When \code{TRUE}, only minimal information is
extracted, including read id, pod5 file name, and basecalled sequence.
If \code{FALSE}, a more comprehensive information is extracted, including poly(A)
tail length, coordinates etc.}

\item{cli_log}{Function for logging messages and progress.}
}
\value{
Data frame with columns:
  \describe{
    \item{read_id}{Character. Read identifier}
    \item{sequence}{Character. Basecalled sequence}
    \item{sequence_length}{Integer. Length of the sequence}
    \item{mapping_quality}{Integer. Mapping quality score}
  }
}
\description{
This function extracts various informations from BAM files for reads that
are present in the corresponding Dorado summary file. It returns a data frame
with read IDs and their corresponding additional info, such as basecalled
sequences, poly(A) lengths, poly(A) coordinates etc., which can then be used
in downstream analyses.
}
\examples{
\dontrun{
sequences <- extract_data_from_bam(
  bam_file = "aligned_reads.bam",
  summary_file = "dorado_summary.txt",
  seq_only = TRUE,
  cli_log = message
)
}
}
