% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{handle_discarded_reads}
\alias{handle_discarded_reads}
\title{Classifies a dataframe containing classification of discarded reads.}
\usage{
handle_discarded_reads(analyzed_results_list, nanopolish)
}
\arguments{
\item{analyzed_results_list}{a list object produced by analyze_results function.}

\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}
}
\value{
a dataframe with characteristics of reads excluded from ninetails
analysis. The object contains readnames and classification - column containing
the reason why the particular read was discarded.
}
\description{
The output dataframe includes reads which were not taken into consideration
in ninetails classification pipeline (i.e. reads, which did not fulfill
the analysis requirements).
}
\details{
The output dataframe consists of 2 columns: readname and class. Class column
contains short explanation of main reason why the given read was omitted.
The reason behind this could be (1) insufficient read quality according to the
nanopolish polya function, (2) insufficient polyA tail length, which makes
segmentation and classification impossible, (3) filtering criterion applied
by the user (reads marked by nanopolish as "SUFFCLIP" may be included/excluded,
depending on user's preference).
}
\examples{
\dontrun{

handle_discarded_reads(analyzed_results_list = list_object,
                       nanopolish = "path/to/nanopolish.tsv/file")

}

}
