% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{is_multifast5}
\alias{is_multifast5}
\title{Check if fast5 file is multi-read format}
\usage{
is_multifast5(fast5_file_structure)
}
\arguments{
\item{fast5_file_structure}{A data frame returned by rhdf5::h5ls() containing
the HDF5 structure of a fast5 file. Must contain a 'name' column.}
}
\value{
Logical. TRUE if the file is multi-read fast5 format, FALSE otherwise.
}
\description{
Checks whether the provided fast5 file structure corresponds to a multi-read
fast5 file format. Multi-read fast5 files contain multiple reads stored
in groups named with "read_" prefix.
}
\examples{
\dontrun{
# Check if a fast5 file is multi-read format
fast5_path <- "path/to/file.fast5"
file_structure <- rhdf5::h5ls(fast5_path, recursive = FALSE)
is_multifast5(file_structure)
}

}
\keyword{internal}
