% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{is_string}
\alias{is_string}
\title{Test if x is a single non-empty character string}
\usage{
is_string(x, min.chars = 1, null.ok = FALSE)
}
\arguments{
\item{x}{Object to test}

\item{min.chars}{Minimum number of characters (default 1)}

\item{null.ok}{Allow NULL? (default FALSE)}
}
\value{
TRUE if x is a valid string, FALSE otherwise
}
\description{
Base R replacement for checkmate::test_string
}
\keyword{internal}
