% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{nanopolish_qc}
\alias{nanopolish_qc}
\title{Aggregates the quality control info produced by nanopolish polya function.}
\usage{
nanopolish_qc(class_data, grouping_factor = NA)
}
\arguments{
\item{class_data}{A dataframe or tibble containig class_data output
from Ninetails

CAUTION! Do not use \code{\link{merge_nonA_tables}} output, since it is cleaned from
poor quality reads by default!}

\item{grouping_factor}{[character string] variable used for grouping the data
(e.g. by sample_name)}
}
\value{
A tibble with counts for each qc_tag present in the run
}
\description{
This function returns read counts assigned to each of the qc_tags
per user-predefined grouping variable (grouping_factor) which might be either
a sample name or an experiment condition (the column of choice must be present
within the input table).
}
\details{
This is the ninetails' implementation of
\code{\link[nanotail:get_nanopolish_processing_info]{name}}
function originally written by P. Krawczyk (smeagol) and incorporated within
the NanoTail package.

For original source code, see:
https://github.com/LRB-IIMCB/nanotail/blob/master/R/polya_stats.R

The variable names were adjusted according to the naming convention within
ninetails to avoid confusion.

Many thanks to the author of original source code for help and advice.
}
