% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{no_na}
\alias{no_na}
\title{Check for no NA values}
\usage{
no_na(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
TRUE if no NA values present, FALSE otherwise
}
\description{
Base R replacement for assertthat::noNA
}
\keyword{internal}
