% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_class_counts}
\alias{plot_class_counts}
\title{Plotting read classes data per category assigned to the analyzed reads.}
\usage{
plot_class_counts(
  class_data,
  grouping_factor = NA,
  frequency = TRUE,
  detailed = TRUE
)
}
\arguments{
\item{class_data}{A dataframe or tibble containing read_classes predictions
made by ninetails pipeline}

\item{grouping_factor}{character string. A grouping variable (e.g. "sample_name")}

\item{frequency}{logical [TRUE/FALSE]. If TRUE, the frequency will be plotted.
If FALSE, raw counts will be shown. This parameter is set to TRUE by default.}

\item{detailed}{logical [TRUE/FALSE]. If TRUE, the counts will be provided
based on the "comments" column, which contains detailed information on the
assigned class. If FALSE, the counts will be provided based on "class" column
which gives more crude glimpse on the classification - i.e. provides an info
whether the reads were considered as "modified", "unmodified" and
"unclassified" only. By default, the TRUE option is set.}
}
\value{
ggplot object with read class prediction
}
\description{
This function requires as input the dataframe with read_classes provided by
ninetails pipeline. Function works in 2 flavours: it can output detailed
classification (based on column "comments") or crude classification (based on
column "class").
}
\details{
Function  based on the Nanotail equivalent: https://github.com/LRB-IIMCB/nanotail
Many thanks to smeagol for advice & support!
}
\examples{
\dontrun{

ninetails::plot_class_counts(class_data=read_classes_dataframe,
                             grouping_factor="sample_name",
                             frequency=TRUE,
                             detailed=TRUE)
}


}
