% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_gasf}
\alias{plot_gasf}
\title{Creates a visual representation of gramian angular summation field
corresponding to the given poly(A) tail fragment (chunk).}
\usage{
plot_gasf(gasf_name, gasf_list, save_file = FALSE)
}
\arguments{
\item{gasf_name}{character string. Name of the given read segment (chunk)
for which the gasf is meant to be plotted. This is the name of the given gasf
within the gasf_list produced by the create_gasf_list function.}

\item{gasf_list}{A list of gasf matrices organized by the read ID_index.}

\item{save_file}{logical [TRUE/FALSE]. If TRUE, the gasf plot 100x100 (pixels)
would be saved in the current workng directory. If FALSE, the plot would be
displayed only in the window. This parameter is set to FALSE by default.}
}
\value{
gramian angular summation field representing given fragment
of nanopore read signal.
}
\description{
The function uses a rainbow palette from grDevices, which makes the matrix
more visually pleasing then greyscale. However, please keep in mind that the
matrices are originally one-channel (computer "sees" them as greyscale).
}
\details{
User can control the size of the plot by defining the dimensions within the
code chunk in R/RStudio. However, please keep in mind that the ninetails'
default built-in model was trained on 100x100 gasfs.
}
\examples{
\dontrun{

plot_gasf(gasf_name = "n4m3-of-giv3n-r3ad_11",
          gasf_list = list_of_gasfs,
          save_file = FALSE)

}
}
