% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_multiple_gaf}
\alias{plot_multiple_gaf}
\title{Creates a visual representation of multiple gramian angular fields
based on provided gaf_list (plots all gafs from the given list).}
\usage{
plot_multiple_gaf(gaf_list, num_cores)
}
\arguments{
\item{gaf_list}{A list of gaf matrices organized by the read ID_index.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
multiple png files containing gramian angular fields representing
given fragment of nanopore read signal.
}
\description{
The function saves the plots to files with predefined size 100x100 pixels.
}
\details{
A feature useful for creating custom data sets for network training.
It is recommended to use this feature with caution. Producing multiple graphs
from extensive data sets may cause the system to crash.

This function plots one-channel (100,100,1) as well as
multi-channel gafs (e.g. 100,100,2).

IMPORTANT NOTE! In current version, this function plots multi-channel matrices
in collapsed manner. If one wants to separate color spaces to GASF/GADF
channels or to R,G,B space, this function would not be suitable. The data
would require additional processing steps!
}
\examples{
\dontrun{

plot_multiple_gaf(gaf_list = list_of_gafs,
                   num_cores = 4)

}
}
