% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_multiple_gasf}
\alias{plot_multiple_gasf}
\title{Creates a visual representation of multiple gramian angular summation fields
based on provided gasf_list (plots all gasfs from the given list).}
\usage{
plot_multiple_gasf(gasf_list, num_cores)
}
\arguments{
\item{gasf_list}{A list of gasf matrices organized by the read ID_index.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
gramian angular summation field representing given fragment
of nanopore read signal.
}
\description{
The function saves the plots to files with predefined size 100x100 pixels.
}
\details{
A feature useful for creating custom data sets for network training.
It is recommended to use this feature with caution. Producing multiple graphs
from extensive data sets may cause the system to crash.
}
\examples{
\dontrun{

plot_multiple_gasf(gasf_list = list_of_gasfs,
                   num_cores = 4)

}
}
