% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions_guppy.R
\name{plot_nonA_abundance}
\alias{plot_nonA_abundance}
\title{Plot abundances of reads with given amount of non-A residues per read}
\usage{
plot_nonA_abundance(residue_data, grouping_factor = NA)
}
\arguments{
\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline}

\item{grouping_factor}{character string. A grouping variable (e.g. "sample_name")}
}
\value{
ggplot2 object
}
\description{
This function plots frequencies of reads containing one, two or more separate
instances (occurrences) of non-As reported by ninetails. The frequency
is computed with respect to the total amount of decorated reads in the
analyzed dataset.
}
\examples{
\dontrun{
plt <- ninetails::plot_nonA_abundance(residue_data = residue_data,
                                      grouping_factor = "sample_name")
plt
}
}
