% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_residue_counts}
\alias{plot_residue_counts}
\title{Plot counts of nonadenosine residues found in ninetails output data}
\usage{
plot_residue_counts(residue_data, grouping_factor = NA, frequency = TRUE)
}
\arguments{
\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline}

\item{grouping_factor}{grouping variable (e.g. "sample_name")}

\item{frequency}{logical [TRUE/FALSE]. If TRUE, the frequency will be plotted.
If FALSE, raw counts will be shown. This parameter is set to TRUE by default.}
}
\value{
a ggplot2 object
}
\description{
Plot counts of nonadenosine residues found in ninetails output data
}
\examples{
\dontrun{
ninetails::plot_residue_counts(residue_data=nonadenosine_residues_dataframe,
                               grouping_factor="sample_name",
                               frequency=TRUE)
}
}
