% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions_guppy.R
\name{plot_rug_density}
\alias{plot_rug_density}
\title{Scatterplot of nonA residue positions within poly(A) tail}
\usage{
plot_rug_density(residue_data, base, max_length)
}
\arguments{
\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline.}

\item{base}{character. One of the following ["C"/"G","U"]. This parameter
defines which nonadenosine is to be plotted. It is obligatory to prevent
the overplotting.}

\item{max_length}{numeric [1]. This parameter controls maximum length
of the poly(A) tail to be taken into consideration for plotting.}
}
\value{
a ggplot2 object
}
\description{
This function allows to produce the scatterplot of raw non-A
residue predictions (y-axis) along the user-defined tail span
(x-axis). The resulting plot also contains the rugs and ridges
along the axes, which are a visual aid designed to better
show the distributions of given residue.
}
\examples{
\dontrun{
ninetails::plot_rug_density(residue_data=residue_data, base="C", max_length=100)
}

}
