% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions.R
\name{plot_tail_range}
\alias{plot_tail_range}
\title{Draws tail range squiggle for given read}
\usage{
plot_tail_range(
  readname,
  nanopolish,
  sequencing_summary,
  workspace,
  basecall_group = "Basecall_1D_000",
  moves = FALSE
)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{basecall_group}{character string ("Basecall_1D_000" is set
as a default). Name of the level in the Fast5 file hierarchy from
which the data should be extracted.}

\item{moves}{logical [TRUE/FALSE]. If TRUE, moves would be plotted in
the background as vertical bars/gaps (for values 0/1, respectively)
and the signal (squiggle) would be plotted in the foreground.
Otherwise, only the signal would be plotted. As a default,
"FALSE" value is set.}
}
\value{
ggplot2 object with squiggle plot
}
\description{
Creates segmented plot of raw ONT RNA signal (with or without moves).
Moves may be plotted only for reads basecalled by Guppy basecaller.
A standalone function; does not rely on any other preprocessing,
depends solely on Nanopolish, Guppy and fast5 input.
}
\examples{
\dontrun{

plot_tail_range(nanopolish = '/path/to/file',
                sequencing_summary = '/path/to/file',
                workspace = '/path/to/guppy/workspace',
                basecalled_group = 'Basecall_1D_000',
                num_cores = 3, moves=TRUE)

}
}
