% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_classes.R
\name{predict_classes}
\alias{predict_classes}
\title{Classification of matrices produced from read chunks with CNN.}
\usage{
predict_classes(gasf_list)
}
\arguments{
\item{gasf_list}{[list] A list of gasf matrices organized by the read ID_index.}
}
\value{
a list of gasfs predictions based on used model.
}
\description{
Classification of matrices produced from read chunks with CNN.
}
\examples{
\dontrun{

predict_classes(gasf_list = gasf_list, keras_model = "/path/to/the_model.h5")

}
}
