% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dorado_input.R
\name{preprocess_dorado_input}
\alias{preprocess_dorado_input}
\title{Preprocess Dorado input files for poly(A) analysis}
\usage{
preprocess_dorado_input(
  bam_file,
  dorado_summary,
  pod5_dir,
  num_cores,
  qc,
  save_dir,
  prefix,
  part_size,
  cli_log
)
}
\arguments{
\item{bam_file}{Path to the BAM file containing aligned reads}

\item{dorado_summary}{Path to Dorado summary file or a data frame containing summary information}

\item{pod5_dir}{Directory containing pod5 files}

\item{num_cores}{Number of CPU cores to use for processing}

\item{qc}{Logical indicating whether to perform quality control checks}

\item{save_dir}{Directory where output files will be saved}

\item{prefix}{Prefix to add to output file names (optional)}

\item{part_size}{Number of reads to process in each chunk}

\item{cli_log}{Function for logging messages and progress}
}
\value{
A list containing paths to processed files:
  \itemize{
    \item summary_files - Paths to split summary files
    \item bam_files - Paths to split BAM files
    \item polya_files - Paths to extracted poly(A) data files
  }
}
\description{
Preprocess Dorado input files for poly(A) analysis
}
\examples{
\dontrun{
processed_files <- preprocess_dorado_input(
  bam_file = "path/to/aligned.bam",
  dorado_summary = "path/to/summary.txt",
  pod5_dir = "path/to/pod5/",
  num_cores = 2,
  qc = TRUE,
  save_dir = "path/to/output/",
  prefix = "experiment1",
  part_size = 40000
)
}
}
