% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dorado_summary.R
\name{process_dorado_summary}
\alias{process_dorado_summary}
\title{Process and optionally split dorado summary file}
\usage{
process_dorado_summary(
  dorado_summary,
  save_dir,
  part_size = 1e+05,
  cli_log = message
)
}
\arguments{
\item{dorado_summary}{character string or data.frame. Either full path of the .txt file
with dorado summary or an in-memory data.frame containing dorado summary data.}

\item{save_dir}{character string. Full path of the directory where the processed
files should be stored.}

\item{part_size}{numeric [100000] (optional). If provided, defines maximum
number of reads per output part.}

\item{cli_log}{Function for logging. This function is encoded in main
pipeline wrapper. Its purpose is to provide neatly formatted & informative log file.}
}
\value{
A character vector containing paths to the generated files.
}
\description{
This function processes a dorado alignment summary file. If the file contains more
reads than the specified part_size, it splits it into multiple smaller files.
Otherwise, it creates a copy in the summary directory. All files are saved in
'dorado_summary_dir' within the specified save directory.
}
\details{
The 'dorado_summary_dir' folder structure:
- If input size <= part_size: Contains a single file named 'dorado_summary.txt'
- If input size > part_size: Contains multiple files named 'part_XX.txt' where XX
  is a zero-padded number indicating the part number
}
