% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{read_residue_multiple}
\alias{read_residue_multiple}
\title{Reads multiple ninetails nonadenosine_residues outputs at once}
\usage{
read_residue_multiple(samples_table, ...)
}
\arguments{
\item{samples_table}{data.frame or tibble containing samples metadata and
paths to files.
It should contain at least two columns: \itemize{
\item residue_path - containing path to the nonadenosine_residues predictions file
\item sample_name - unique sample identifier
}}

\item{...}{- additional parameters to pass to read_residue_single();
currently accepts sample_name argument

Function based on PK (smaegol) NanoTail read_polya_multiple
https://github.com/LRB-IIMCB/nanotail/
#' Many thanks to the NanoTail developer for help and advice!}
}
\value{
a [tibble] containing non-A residue data for
all specified samples, with metadata provided in samples_table
stored as separate columns
}
\description{
This function can be used to load any number of files
with nonadenosine_residues predictions with single invocation,
allowing for metadata specification.
}
\details{
It requires providing a samples_table with metadata necessary
to identify each sample.
}
\examples{
\dontrun{

samples_table <- data.frame(residue_path = c(path_1,path_2,path_3,path_4),
                            sample_name =c("wt_1","mut_1","wt_2","mut_2"),
                            group = c("wt","mut","wt","mut"))

residues_data <- ninetails::read_residue_multiple(samples_table)

}
}
