% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_gasf.R
\name{resize_gasf}
\alias{resize_gasf}
\title{Resizes gramian angular summation field.}
\usage{
resize_gasf(gasf, new_dims = dim(gasf))
}
\arguments{
\item{gasf}{- an array, gramian angular summation matrix produced by
create_gasf() function.}

\item{new_dims}{numeric vector. Dimensions of new matrix.}
}
\value{
resized gasf (according to defined dims)
}
\description{
Resizes gramian angular summation field.
}
\examples{
\dontrun{
resize_gasf <- function(gasf=gasf, new_dims=c(100,100,3))
}
}
