% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_guppy.R
\name{save_outputs}
\alias{save_outputs}
\title{Save pipeline outputs to files}
\usage{
save_outputs(outputs, save_dir, prefix = "")
}
\arguments{
\item{outputs}{List containing read_classes and nonadenosine_residues data frames}

\item{save_dir}{Directory to save the files}

\item{prefix}{Optional prefix for file names}
}
\value{
Invisible NULL
}
\description{
Save pipeline outputs to files
}
