% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bam_file.R
\name{split_bam_file}
\alias{split_bam_file}
\title{Split BAM file into parts based on readnames from dorado summary}
\usage{
split_bam_file(
  bam_file,
  dorado_summary,
  part_size = 1e+05,
  save_dir,
  cli_log = message
)
}
\arguments{
\item{bam_file}{character string. Full path to the input BAM file.}

\item{dorado_summary}{character string. Full path to the dorado summary part file.}

\item{part_size}{numeric [100000] (optional). If provided, defines maximum
number of reads per output part.}

\item{save_dir}{character string. Full path of the directory where the processed
BAM files should be stored.}

\item{cli_log}{Function for logging. This function is encoded in main
pipeline wrapper. Its purpose is to detailed log file.}
}
\value{
A character vector containing paths to the generated BAM files.
}
\description{
This function splits a large BAM file into multiple smaller BAM files based on
unique read identifiers (readnames) from a dorado alignment summary file parts.
}
\examples{
\dontrun{

# Split a BAM file based on dorado summary part
split_bam_file(
  bam_file = '/path/to/input.bam',
  dorado_summary = '/path/to/summary_part.txt',
  save_dir = '/path/to/output',
  part_size = 1000
)

}
}
