% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_guppy.R
\name{split_polya_data}
\alias{split_polya_data}
\title{Split large poly(A) data file into smaller polya_parts}
\usage{
split_polya_data(polya_data, part_size = 1e+05, save_dir)
}
\arguments{
\item{polya_data}{Data frame or path to file containing poly(A) data}

\item{part_size}{Maximum number of rows per polya_part}

\item{save_dir}{Directory to save polya_parts}
}
\value{
List of file paths to created polya_parts
}
\description{
Split large poly(A) data file into smaller polya_parts
}
