% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_with_overlaps_moved.R
\name{split_with_overlaps_moved}
\alias{split_with_overlaps_moved}
\title{Creates vector of overlapping tail fragments extracted from
nanopolish output and fast5 files, in which move value = 1 occurs.}
\usage{
split_with_overlaps_moved(readname, tail_feature_list, segment, overlap)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{segment}{numeric [1]. Length of the chunk(s) to be created.}

\item{overlap}{numeric [1]. Length of the overlap between the chunks.}
}
\value{
A vector of indices of moved chunks organized by their order in
an original list is returned.
}
\description{
Creates vector of overlapping tail fragments extracted from
nanopolish output and fast5 files, in which move value = 1 occurs.
}
\examples{
\dontrun{

split_with_overlaps_moved(readname = "name_of_given_read_of_interest",
                          tail_feature_list = list_of_tail_features,
                          segment = 100
                          overlap = 10)

}


}
