% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{substitute_gaps}
\alias{substitute_gaps}
\title{Substitutes 0s surrounded by adjacent nonzeros in pseudomove vector
to facilitate position-centering function.}
\usage{
substitute_gaps(pseudomoves)
}
\arguments{
\item{pseudomoves}{numeric vector produced by z-score
filtering algo (filter_signal_by_threshold() function) corresponding
to the tail region of the read of interest as delimited by
nanopolish polya function.}
}
\value{
a numeric vector of adjusted pseudomoves (smoothened)
}
\description{
This function helps to avoid redundancy introduced by z-score thesholding
algo (this happens when signal is jagged), so one segment would be reported
instead of multiple.
}
\examples{
\dontrun{

substitute_gaps(pseudomoves = pseudomoves_vector)

}

}
