% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_squiggle.R
\name{view_squiggle}
\alias{view_squiggle}
\title{Creates segmented plot of given raw ONT RNA signal.}
\usage{
view_squiggle(
  readname,
  nanopolish,
  sequencing_summary,
  workspace,
  basecall_group = "Basecall_1D_000",
  num_cores
)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{basecall_group}{character string ("Basecall_1D_000" is set
as a default). Name of the level in the Fast5 file hierarchy from
which the data should be extracted.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{

}
\description{
Creates segmented plot of given raw ONT RNA signal.
}
