% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{winsorize_signal}
\alias{winsorize_signal}
\title{Winsorizes nanopore signal - removes high cliffs (extending above & below
signal range). Slightly affects signal extremes.}
\usage{
winsorize_signal(signal)
}
\arguments{
\item{signal}{character string. Name of the given ONT signal.}
}
\value{
winsorized signal.
}
\description{
Based on the A. Signorell's DescTools
https://github.com/AndriSignorell/DescTools/blob/master/R/DescTools.r
}
\examples{
\dontrun{

winsorize_signal <- function(signal='12fdcb3-ewfd543-34552-1234ddta345')

}

}
