% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{analyze_results}
\alias{analyze_results}
\title{Summarizes classification results into list of 2 dataframes containing two
types of outputs (read type classification and detailed positional info).}
\usage{
analyze_results(nanopolish, coordinate_df, predicted_list, pass_only = TRUE)
}
\arguments{
\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{coordinate_df}{a data frame object produced by create_coordinate_df
function.}

\item{predicted_list}{a list object produced by predict_classes function.}

\item{pass_only}{logical [TRUE/FALSE]. This must be consistent with the
parameter set while using previous functions requiring that option in an
entire analysis. This means, if in previously used create_tail_feature_list()
function parameter pass_only was set to "FALSE" it shall be set to "FALSE" in
analyze_results() function either. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}
}
\value{
a list of 2 dataframes with prediction results. First dataframe
contains initial indications, whether the given read was recognized as
modified (containing non-adenosine residue) or not.  The second dataframe
contains detailed info on type and estimated positions of non-adenosine
residues detected.
}
\description{
First dataframe contains classification of ONT sequencing reads based on
their type. This means that reads are assigned to one of the following
classes: [1] "modified" - if they meet initial conditions for analysis and
contain potential non-Adenosine residue, [2] "unmodified" - if they contain
no move =1 (only 0s) or contain move =1 but were classified as "A" containing
exclusively by neural network, [3] "unclassified - insufficient read length"
if their length estimated by nanopolish polya function was lower or equal
to 10 nt, [4] other unclassified cases which may appear in the dataset were
excluded from the analysis due to their nanopolish qc tag - the reason for
this is given within the comment.
}
\details{
Second dataframe contains detailed positions of detected modifications.
}
\examples{
\dontrun{

analyze_results(nanopolish = '/path/to/nanopolish_output.tsv',
                coordinate_df = '/path/to/coordinate_df',
                predicted_list = '/path/to/predicted_list',
                pass_only=TRUE)

}

}
