% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_annotation.R
\name{annotate_with_biomart}
\alias{annotate_with_biomart}
\title{Annotates the ninetails output data}
\usage{
annotate_with_biomart(
  input_data,
  attributes_to_get = c("ensembl_transcript_id", "external_gene_name", "description",
    "transcript_biotype"),
  filters = "ensembl_transcript_id",
  organism = NULL,
  mart_to_use = NULL
)
}
\arguments{
\item{input_data}{[dataframe] tabular output of ninetails pipeline.}

\item{attributes_to_get}{[character] annotation attributes to be retrieved.
By default: 'external_gene_name','description','transcript_biotype'.}

\item{filters}{column of the input dataframe to be matched
with the target mart (e.g. "ensembl_transcript_id_short")}

\item{organism}{[character] character string. Currently available::\itemize{
\item 'athaliana' - Arabidopsis thaliana
\item 'hsapiens' - Homo sapiens
\item 'mmusculus' - Mus musculus
\item 'scerevisiae' - Saccharomyces cerevisiae
}
This argument is optional & mutually exclusive with 'mart_to_use'.}

\item{mart_to_use}{optional: mart object created with \link[biomaRt]{useMart}
or \link[biomaRt]{useEnsembl} function. This argument is optional
& mutually exclusive with 'organism'.}
}
\value{
A dataframe with annotated ninetails output data.
}
\description{
This function is a wrapper for data annotation using the biomaRt package.
}
\details{
For proper execution, this function requires the biomaRt package version
greater than 2.40.

One of the arguments 'organism' or 'mart_to_use' must be defined.
The arguments 'organism' and 'mart_to_use' are mutually exclusive.
If both are declared, the function will throw an error.

Function based on PK (smaegol) NanoTail annotate_with_biomart()
https://github.com/LRB-IIMCB/nanotail/
Many thanks to the NanoTail developer for help and kind advice!
}
\examples{
 \dontrun{
# with provided organism
annot <- ninetails::annotate_with_biomart(input_data=merged_nonA_tables,
                                          organism="mmusculus")

# with provided mart (where 'mart' is the output of \link[biomaRt]{useMart} function)
annot <- ninetails::annotate_with_biomart(input_data=merged_nonA_tables,
                                          mart_to_use=mart)
}

}
