% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{check_fast5_filetype}
\alias{check_fast5_filetype}
\title{Checks if the provided directory contains fast5 files in the correct format.}
\usage{
check_fast5_filetype(workspace, basecall_group)
}
\arguments{
\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the Fast5 file hierarchy from which the data should be extracted.}
}
\value{
outputs the text info with basic characteristics of the data.
}
\description{
This function analyses the structure of the first fast5 file in the given
directory and checks whether it fulfills the analysis requirements (if the
file is multifast5, basecalled by Guppy basecaller and containing provided
basecall_group). Otherwise the function throws an error (with description).
}
\examples{
\dontrun{

check_fast5_filetype <- function(workspace = '/path/to/guppy/workspace',
                                 basecalled_group = 'Basecall_1D_000')

}



}
