% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tails.R
\name{check_tails}
\alias{check_tails}
\title{Creating parameters for nnet.}
\usage{
check_tails(
  nanopolish,
  sequencing_summary,
  workspace,
  num_cores,
  basecall_group,
  pass_only
)
}
\arguments{
\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the Fast5 file hierarchy from which the data should be extracted.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}
}
\value{
A list containing read information organized by the read ID
is returned. Always assign this returned list to a variable, otherwise
the long list will be printed to the console, which may crash your R session.
}
\description{
Creating parameters for nnet.
}
\examples{
\dontrun{

check_tails(nanopolish = '/path/to/file',
            sequencing_summary = '/path/to/file',
            workspace = '/path/to/guppy/workspace',
            num_cores = 10,
            basecalled_group = 'Basecall_1D_000',
            pass_only=TRUE)

}
}
