% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_dorado_DRS.R
\name{check_tails_dorado_DRS}
\alias{check_tails_dorado_DRS}
\title{Wrapper function for complete DRS processing by ninetails package (POD5/Dorado mode).}
\usage{
check_tails_dorado_DRS(
  bam_file,
  dorado_summary,
  pod5_dir,
  num_cores = 1,
  qc = TRUE,
  save_dir,
  prefix = "",
  part_size = 1000
)
}
\arguments{
\item{bam_file}{character string. Full path to the BAM file produced by
Dorado's align function. The BAM file must be sorted and indexed.}

\item{dorado_summary}{character string or data.frame. Either full path of the .txt file
with dorado summary or an in-memory data.frame containing dorado summary data.}

\item{pod5_dir}{character string. Full path of the directory containing
POD5 files. The POD5 files must correspond to the reads in the BAM file.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.
This parameter is set to 1 by default.}

\item{qc}{logical [TRUE/FALSE]. If TRUE, quality control of nonadenosine
residue positions will be performed. As a default, "TRUE" value is set.}

\item{save_dir}{character string. Full path of the directory where the output
files containing the tail composition information should be stored.}

\item{prefix}{character string (optional). If provided, it will be inserted
into the names of the output files, positioned between the timestamp
and the suffix indicating the file type.}

\item{part_size}{numeric [100000] (optional). If provided, defines maximum
number of reads to process in a single batch. Default is 100000 reads.}
}
\value{
A list containing tail information organized by the read ID.
Additionally, creates timestamped read_classes & nonadenosine_residues
files in the specified output directory.
}
\description{
This function processes direct RNA sequencing (DRS) data generated using
R10 chemistry SQK-RNA004 flowcells and analyzed with Dorado basecaller.
It accepts BAM files produced by Dorado's align function and POD5 files
containing raw signal data.
}
\details{
This function allows to perform all of the steps required to discover
nonadenosine nucleotides within poly(A) tails using raw signal analysis.
Please keep in mind, that during computations the function creates large
segmentation data. Therefore it may be wise to process data in parts,
which is handled automatically by the function.

The output of this function is a list of 2 dataframes, containing:\itemize{
\item read_classes - classification of reads based on applied criteria
\item nonadenosine_residues - detailed positional info regarding all
potential nonadenosine residues detected.
}
}
\examples{
\dontrun{

results <- ninetails::check_tails_dorado_DRS(
  bam_file = '/path/to/aligned.bam',
  pod5_dir = '/path/to/pod5/files',
  dorado_summary = '/path/to/summary.txt',
  num_cores = 2,
  save_dir = '~/Downloads',
  prefix = "prefix",
  part_size = 2000
)

}
}
