% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{correct_residue_data}
\alias{correct_residue_data}
\title{Marks uncertain positions of non-A residues in ninetails output data}
\usage{
correct_residue_data(
  class_data,
  residue_data,
  grouping_factor = NULL,
  transcript_column,
  ref = NULL
)
}
\arguments{
\item{class_data}{[dataframe] A dataframe or tibble containing read_classes
predictions made by ninetails pipeline}

\item{residue_data}{[dataframe] A dataframe or tibble containig non-A residue
predictions made by ninetails pipeline}

\item{grouping_factor}{[character string]. A grouping variable (e.g. "sample_name",
"group", etc.) - optional.}

\item{transcript_column}{[character string]. Name of column containing transcript id
(e.g. "ensembl_transcript_id_short") - required.}

\item{ref}{[character string] or object or NULL (default) - the whitelist of transcripts
with hybrid tails - containing 3'UTRs highly enriched with A nucleotides (in last 20
positions >80%).
Current version of ninetails contains built-in whitelists, which may be selected
by the user:\itemize{
\item 'athaliana' - Arabidopsis thaliana
\item 'hsapiens' - Homo sapiens
\item 'mmusculus' - Mus musculus
\item 'scerevisiae' - Saccharomyces cerevisiae
\item 'celegans' - Caenorhabditis elegans
\item 'tbrucei' - Trypanosoma brucei
}
It is also possible to provide own whitelist. Important note: whitelist must be consistent
with the content of the 'transcript_column'. Using whitelist is not mandatory,
however it allows to retrieve more true positive data.}
}
\value{
A tibble with decorated residue_data. In this dataframe, in comparison
to the original (raw) residue_data, some additional columns are present:\itemize{
\item mode_pos - most frequent position of non-A reported for given transcript
\item mode_len - most frequent tail length reported for given transcript
\item seg_err_quart - 0.05 quantile of tail length for given transcript
\item qc_pos - quality tag for given position "Y" for correct, "N" for ambiguous
\item pos_err_quart - 0.05 quantile of given nonA (C, G, U, respectively) for given transcript
\item count_nonA - number of nonA containing reads for given transcript
\item count - number of reads for given transcript
}
}
\description{
Based on quantiles of poly(A) tail length & non-A residue distribution
peak calling per transcript.
}
\examples{
\dontrun{
# the output of function \code{\link{check_tails_guppy}} or \code{\link{create_outputs}}
# is required (below denoted as "results") to run the following example
# for details see the documentation for the \code{\link{check_tails_guppy}}
# and \code{\link{create_outputs}}

residue_data_edited <- ninetails::correct_residue_data(class_data=results[[1]],
                                                       residue_data=results[[2]],
                                                       transcript_column="contig")
}

}
