% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gasf.R
\name{create_gasf}
\alias{create_gasf}
\title{Converting ONT signal to gramian angular summation field.}
\usage{
create_gasf(chunkname, tail_chunk_list)
}
\arguments{
\item{chunkname}{character string. Name of the given signal chunk within the
analyzed dataset.}

\item{tail_chunk_list}{list object produced by create_tail_chunk_list function.}
}
\value{
an array (100,100,1) with values (greyscale) representing ONT signal.
}
\description{
Converting ONT signal to gramian angular summation field.
}
\examples{
\dontrun{

create_gasf(chunkname = '1234-jhgfdr54-io643gju-ouy4378989', tail_chunk_list = tail_chunk_list)

}



}
