% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ninetails_training_dataset_production_functions.R
\name{create_tail_chunk_list_A}
\alias{create_tail_chunk_list_A}
\title{Creates list of tail chunks containing As exclusively.}
\usage{
create_tail_chunk_list_A(tail_feature_list, num_cores)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
a list object (nested) containing the segmented tail data  (chunks,
coordinates) organized by the read IDs.
}
\description{
Extracts fragments of polyA tails of ONT RNA reads containing only A
nucleotides along their coordinates & appends the data to the nested
list organized by read IDs.
}
\details{
This version of the function is intended to be used to produce training and
validation datasets. It is designed to filter signal fragments that DO NOT
contain potential non-A nucleotides.

This function segments the signals with overlaps, making the data multiplied (data augmentation).
}
\examples{
\dontrun{

create_tail_chunk_list_A(tail_feature_list = tail_feature_list,
                         num_cores = 2)

}

}
