% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ninetails_training_dataset_production_functions.R
\name{create_tail_chunk_list_trainingset}
\alias{create_tail_chunk_list_trainingset}
\title{Extracts fragments of polyA tails of ONT RNA reads containing non-A
nucleotides along their coordinates & appends the data to the nested
list organized by read IDs.}
\usage{
create_tail_chunk_list_trainingset(tail_feature_list, num_cores)
}
\arguments{
\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.}
}
\value{
a list object 9nested) containing the segmented tail data  (chunks,
coordinates) organized by the read IDs.
}
\description{
This version of the function is intended to be used to produce training and
validation datasets.
}
\examples{
\dontrun{

create_tail_chunk_list_trainingset(tail_feature_list = tail_feature_list,
                                   num_cores = 2)
}
}
