% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{extract_polya_data}
\alias{extract_polya_data}
\title{Extracts features of multiple RNA reads from nanopolish output
& sequencing summary file.}
\usage{
extract_polya_data(nanopolish, sequencing_summary, pass_only = TRUE)
}
\arguments{
\item{nanopolish}{character string. Full path of the .tsv file produced
by nanopolish polya function.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}
}
\value{
A tibble/df containing read information organized by the read ID
is returned. Always assign this returned tibble to a variable. Otherwise
the long tibble will be printed to the console, which may crash your
R session.
}
\description{
This function extracts features of poly(A) tails of selected RNA reads
from the output table provided by nanopolish polya function and sequencing
summary provided by the sequencer. Filenames are taken from the sequencing
summary file. Only reads with tail lengths estimated as >= 10nt by nanopolish
polya function are taken into account.
}
\examples{
\dontrun{

extract_polya_data(nanopolish = '/path/to/nanopolish/polya/output.tsv',
                   sequencing_summary = '/path/to/sequencing_summary.txt',
                   pass_only = TRUE)

}


}
