% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{extract_tail_data}
\alias{extract_tail_data}
\title{Extracts tail features of single RNA read from respective basecalled
multi-fast5 file.}
\usage{
extract_tail_data(readname, polya_summary, workspace, basecall_group)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{polya_summary}{character string. The table containing data extracted
from nanopolish & sequencing summary (using extract_polya_data() function.}

\item{workspace}{character string. Full path of the directory to search
the basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the fast5 file hierarchy from which the data should be extracted.}
}
\value{
A list containing read information organized by the read ID
is returned. Always assign this returned list to a variable, otherwise
the long list will be printed to the console, which may crash your R session.
}
\description{
This function extracts metadata of RNA read from multi-fast5 file basecalled
by guppy. Filenames are taken from the sequencing
summary file. Tail signal (as delimited by nanopolish polya function) is
extracted (before the extraction step, the entire signal is winsorized
so the cliffs are removed), and downsampled (to facilitate further analysis).
}
\examples{
\dontrun{

ninetails::extract_tail_data(readname = 'abc123de-fg45-6789-0987-6543hijk2109',
                             polya_summary = polya_summary_table,
                             workspace = '/path/to/folder/containing/multifast5s',
                             basecall_group = 'Basecall_1D_000')

}
}
