% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{filter_signal_by_threshold}
\alias{filter_signal_by_threshold}
\title{Detection of outliers (peaks & valleys) in ONT signal using z-scores.}
\usage{
filter_signal_by_threshold(signal)
}
\arguments{
\item{signal}{numeric vector. An ONT read fragment corresponding to the tail
region of the read of interest as delimited by nanopolish polya function (the
fragments are stored in tail_feature_list[[1]] produced by the
\code{\link{create_tail_feature_list}} function.}
}
\value{
numeric vector of "pseudomoves" corresponding to the analyzed tail
region; containing values ranging from -1 to 1.
}
\description{
The function enables detection in the signal corresponding to the poly(A)
tail of areas in which the signal value significantly deviates from the
values typical of an adenosine homopolymer.
}
\details{
The function results in a vector of "pseudomoves" with values in the range
-1:1, where -1 corresponds to signals that are significantly less than the
mean and standard deviation of the typical A-homopolymer, 1 corresponds to
signals that are significantly higher than the mean and standard deviation
of the typical A-homopolymer, and 0 corresponds to the typical values
for the A-homopolymer.

The "pseudomoves" vector allows more accurate calibration of the nucleotide
positions of potential non-adenosine residues than the moves produced
by the guppy basecaller.

The function created based on the following source:
Brakel, J.P.G. van (2014). "Robust peak detection algorithm using z-scores".
Stack Overflow. Available at:
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data/22640362#22640362
(version: 2020-11-08).
}
\examples{
\dontrun{

ninetails::filter_signal_by_threshold(signal=tail_feature_list[[1]][["readname"]][[4]])

}
}
