% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ninetails_training_dataset_production_functions_guppy.R
\name{filter_signal_by_threshold_trainingset}
\alias{filter_signal_by_threshold_trainingset}
\title{Detection of outliers (peaks & valleys) in ONT signal using z-scores.}
\usage{
filter_signal_by_threshold_trainingset(signal)
}
\arguments{
\item{signal}{numeric vector. An ONT read fragment corresponding to the tail
region of the read of interest as delimited by nanopolish polya function (the
fragments are stored in tail_feature_list[[1]] produced by the
create_tail_feature_list() function.}
}
\value{
numeric vector of "pseudomoves" corresponding to the analyzed tail
region; containing values ranging from -1 to 1.
}
\description{
This is the version of the function useful to produce the training set
for CNN. Slightly different from the original one.
}
\details{
The function created based on the following source:
Brakel, J.P.G. van (2014). "Robust peak detection algorithm using z-scores".
Stack Overflow. Available at:
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data/22640362#22640362
(version: 2020-11-08).
}
\examples{
\dontrun{

filter_signal_by_threshold(signal=tail_feature_list[[1]][["readname"]][[4]])

}
}
