% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_keras_model.R
\name{load_keras_model}
\alias{load_keras_model}
\title{Loads keras model for multiclass signal prediction.}
\usage{
load_keras_model(keras_model_path)
}
\arguments{
\item{keras_model_path}{either missing or character string. Full path of the .h5 file
with keras model used to predict signal classes. If function is called without this
argument(argument is missing) the default pretrained model will be loaded. Otherwise,
the dir with custom model shall be provided.}
}
\value{

}
\description{
Loads keras model for multiclass signal prediction.
}
\examples{
\dontrun{

load_keras_model(keras_model_path = "/path/to/the/model/in/hdf5_format")
}

}
