% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{merge_nonA_tables}
\alias{merge_nonA_tables}
\title{Merges ninetails tabular outputs (read classes and nonadenosine residue data)
to produce one concise table for all data.}
\usage{
merge_nonA_tables(class_data, residue_data, pass_only = TRUE)
}
\arguments{
\item{class_data}{A dataframe or tibble containing read_classes predictions
made by ninetails pipeline}

\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.
IMPORTANT NOTE! This option must be set the same as it was during the
pipeline output production.}
}
\value{
a tibble with summarized info from both ninetails outputs
}
\description{
Each read is represented by single row.
The output of this function is a tibble containing additional columns.
The "prediction" column from the residue_data is spread to 3 separate columns
(named with "prediction_" prefix), containing either C, G or U hits per read,
respectively. "Hit" is understood as a single presence of a given non-A residue.
}
\details{
In addition, an extra "nonA_residues" column is located at the end of
the output table. It contains all non-A residues positions summarized (per read),
given from the 5' to 3' end, separated by ":".

In this table, only reads that have been classified by ninetails
are included (reads marked "unclassified" are omitted from the analysis).
}
\examples{
\dontrun{

merged_tables <- ninetails::merge_nonA_tables(class_data=class_data,
                                              residue_data=residue_data,
                                              pass_only=TRUE)

}
}
