% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails-package.R
\docType{package}
\name{ninetails-package}
\alias{ninetails}
\alias{ninetails-package}
\title{ninetails: Nonadenosine Nucleotides in Poly(A) Tails}
\description{
Finds noncanonical (nonadenosine) nucleotides in poly(A) tails of Oxford Nanopore direct RNA sequencing reads using convolutional neural network based on raw current signal.
}
\keyword{internal}
