% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{nonA_fisher}
\alias{nonA_fisher}
\title{Performs Fisher's exact test for testing the null of independence of rows and
columns in a contingency table representing given transcript in ninetails
output data. This is a wrapper for fisher.test function from stats package
with additional features to facilitate data wrangling.}
\usage{
nonA_fisher(
  ninetails_data,
  grouping_factor,
  base,
  min_reads = 0,
  transcript_id_column = NA
)
}
\arguments{
\item{ninetails_data}{dataframe - the output of ninetails::merge_nonA_tables
function (merged tabular output containing read classification &
non-A position data).}

\item{grouping_factor}{[character string] the name of factor variable defining
groups/conditions (needs to have 2 levels!)}

\item{base}{[character string] letter representing particular non-A nucleotide,
for which the statistics are meant to be computed. Currently function accepts
C, G, U arguments. The "C" value is set by default.}

\item{min_reads}{[numeric] minimum number of reads representing given
transcript to include it in the analysis}

\item{transcript_id_column}{[character string] name of the column in which
the identifiers of the transcripts are stored. It is set to NA
by default.}
}
\value{
a tibble with results for given transcript, including pvalue,
adjusted pvalue, stats_code (the variable describing whether conditions
are met) and significance (FDR based on padj).
}
\description{
It is suitable only for the pairwise comparisons (i.e. for 2x2 contingency
table), where 2 conditions (e.g. WT vs KO) are compared at once.
}
\details{
The function allows the user to set a cutoff number of reads required for
the analysis.

This function is intended to work under the calculate_fisher function.

The function was inspired by the Nanotail package written & maintained by
Pawel Krawczyk (smaegol): https://github.com/LRB-IIMCB/nanotail/blob/dev/R/polya_stats.R

Many thanks to the developer of original source code.
}
\examples{
\dontrun{
test <- ninetails::nonA_fisher(ninetails_data=merged_nonA_tables,
                               grouping_factor = "sample_name",
                               base="C",
                               min_reads=100)

}

}
