% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_plotting_functions_guppy.R
\name{plot_panel_characteristics}
\alias{plot_panel_characteristics}
\title{Plots panel characteristics of ninetails output.}
\usage{
plot_panel_characteristics(
  input_residue_data,
  input_class_data = NULL,
  input_merged_nonA_tables_data = NULL,
  type = "default",
  max_length = 300,
  direction_5_prime = TRUE
)
}
\arguments{
\item{input_residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline}

\item{input_class_data}{A dataframe or tibble containing read_classes predictions
made by ninetails pipeline. Mutually exclusive with parameter input_merged_nonA_tables_data}

\item{input_merged_nonA_tables_data}{A dataframe or tibble containing merged_nonA_tables data
produced by the \code{\link{merge_nonA_tables}} function. Mutually exclusive with parameter
input_class_data}

\item{type}{[character] either "default" or "moderna" - controls the predefined
settings of the plot. In case of moderna adds the default UCUAG pentamer position
to the distribution subplots. By default, the "default" option is set.}

\item{max_length}{[numeric] maximum length of plotted tail data}

\item{direction_5_prime}{[logical] by default set to TRUE. controls the direction in which
the non-A positions are reported. Either from 5' end, according to the most common convention,
or from 3' end (as ONT direct RNA seq proceeds).}
}
\value{
a ggplot object
}
\description{
Creates a multipanel plot with comprehensive characteristics of input data
produced by ninetails software. Those panel charts provide the most comprehensive
characterization of a given pool of reads (representing particular transcript
or set of transcripts, respectively).
}
\details{
Includes 5 panels A-E containing various subplots with the following content:\itemize{
\item A - read categories - result of read classification according to presence/absence
of non-As in their poly(A) tails
\item B - non-A residues - frequency of reads containing given residue among all
of the reads harboring non-A nucleotides
\item C - distribution of lengths of poly(A) tails - overall distribution of lengths
in comparison to the lengths of tails decorated with non-As
\item D - normalized distribution of non-A residues in poly(A) tails -
non-A nucleotide positions normalized to the length of reads in which given residue occurs
\item raw distribution of non-A residues in poly(A) tails - crude depiction
of positions along the tail range
}
}
\examples{
\dontrun{

ninetails::plot_panel_characteristics(input_residue_data=residue_data,
                                      input_class_data=class_data,
                                      input_merged_nonA_tables_data=NULL,
                                      type="default",
                                      max_length=100,
                                      direction_5_prime=T)

}

}
