% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_guppy.R
\name{process_polya_complete}
\alias{process_polya_complete}
\title{Variant of Check_tails_guppy processing of unsplitted poly(A) data.}
\usage{
process_polya_complete(
  polya_data,
  sequencing_summary,
  workspace,
  num_cores,
  basecall_group,
  pass_only,
  qc,
  save_dir,
  prefix,
  cli_log
)
}
\arguments{
\item{polya_data}{character string. Full path of the .tsv file produced
by either nanopolish polya function or tailfindr. Only DRS-derived tables are
accepted. In case of tailfindr output, it would be immediately converted to
nanopolish-like format.}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.
This parameter is set to 1 by default.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the Fast5 file hierarchy from which the data should be extracted.
This parameter is set to 'Basecall_1D_000' by default.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}

\item{qc}{logical [TRUE/FALSE]. If TRUE, the quality control of the output
predictions would be performed. This means that the reads/non-A residue
positions in terminal nucleotides, which are most likely artifacts, are
labeled accordingly as "-WARN" (residues recognized as non-A due to
nanopolish segmentation error which is inherited from nanopolish,
as ninetails uses nanopolish segmentation). It is then up to user, whether
they would like to include or discard such reads from their pipeline. However,
it is advised to treat them with caution. By default, the qc option is enabled
(this parameter is set to TRUE).}

\item{save_dir}{character string. Full path of the directory where the output
files containing the tail composition information should be stored.}

\item{prefix}{character string (optional). If provided, it will be inserted
into the names of the output files, positioned between the timestamp
and the suffix indicating the file type (either read classes or
nonadenosine residues).}

\item{cli_log}{Function for logging}
}
\value{
A list containing tail information organized by the read ID
}
\description{
This variant of the pipeline allows to process small to moderate size of
poly(A) data (up to 100000 reads by default).
}
\examples{
\dontrun{
# Process a complete poly(A) dataset
results <- ninetails::process_polya_complete(
  polya_data = "path/to/polya_data.tsv",
  sequencing_summary = "path/to/sequencing_summary.txt",
  workspace = "path/to/fast5_files",
  num_cores = 4,
  basecall_group = "Basecall_1D_000",
  pass_only = TRUE,
  qc = TRUE,
  save_dir = "path/to/output",
  prefix = "experiment1",
  cli_log = cli_log)
}
}
