% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_guppy.R
\name{process_polya_parts}
\alias{process_polya_parts}
\title{Variant of check_tails_guppy processing of poly(A) data splitted to parts.}
\usage{
process_polya_parts(
  part_files,
  sequencing_summary,
  workspace,
  num_cores,
  basecall_group,
  pass_only,
  qc,
  save_dir,
  prefix,
  cli_log
)
}
\arguments{
\item{part_files}{List of file paths to poly(A) data parts}

\item{sequencing_summary}{character string. Full path of the .txt file
with sequencing summary or in-memory object with sequencing summary data.}

\item{workspace}{character string. Full path of the directory to search the
basecalled fast5 files in. The Fast5 files have to be multi-fast5 file.}

\item{num_cores}{numeric [1]. Number of physical cores to use in processing
the data. Do not exceed 1 less than the number of cores at your disposal.
This parameter is set to 1 by default.}

\item{basecall_group}{character string ["Basecall_1D_000"]. Name of the
level in the Fast5 file hierarchy from which the data should be extracted.
This parameter is set to 'Basecall_1D_000' by default.}

\item{pass_only}{logical [TRUE/FALSE]. If TRUE, only reads tagged by
nanopolish as "PASS" would be taken into consideration. Otherwise, reads
tagged as "PASS" & "SUFFCLIP" will be taken into account in analysis.
As a default, "TRUE" value is set.}

\item{qc}{logical [TRUE/FALSE]. If TRUE, the quality control of the output
predictions would be performed. This means that the reads/non-A residue
positions in terminal nucleotides, which are most likely artifacts, are
labeled accordingly as "-WARN" (residues recognized as non-A due to
nanopolish segmentation error which is inherited from nanopolish,
as ninetails uses nanopolish segmentation). It is then up to user, whether
they would like to include or discard such reads from their pipeline. However,
it is advised to treat them with caution. By default, the qc option is enabled
(this parameter is set to TRUE).}

\item{save_dir}{character string. Full path of the directory where the output
files containing the tail composition information should be stored.}

\item{prefix}{character string (optional). If provided, it will be inserted
into the names of the output files, positioned between the timestamp
and the suffix indicating the file type (either read classes or
nonadenosine residues).}

\item{cli_log}{Function for logging}
}
\value{
A list containing tail information organized by the read ID.
}
\description{
In this pipeline variant, the large input poly(A) measurement file is split
to parts and processed sequentially one by one. This is to avoid memory
overflow and facilitate potential debugging.
}
\details{
Creates individual outputs for each part and merges them into final results.
}
\examples{
\dontrun{
test <- ninetails::process_polya_parts(
  part_files=list_of_files,
  sequencing_summary = "/seq/summary/dir/sequencing_summary.txt",
  workspace = "/dir/with/basecalled/fast5/",
  num_cores = 2,
  basecall_group = 'Basecall_1D_000',
  pass_only = TRUE,
  qc = TRUE,
  save_dir = '~/Downloads',
  prefix = "test")
}
}
