% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{read_class_multiple}
\alias{read_class_multiple}
\title{Reads multiple ninetails read_classes outputs at once}
\usage{
read_class_multiple(samples_table, ...)
}
\arguments{
\item{samples_table}{data.frame or tibble containing samples metadata
and paths to files.
Should have at least two columns: \itemize{
\item class_path - containing path to the read_classes predictions file
\item sample_name - unique sample identifier
}}

\item{...}{- additional parameters to pass to read_class_single();
currently accepts sample_name argument

Function based on PK (smaegol) NanoTail read_polya_multiple
https://github.com/LRB-IIMCB/nanotail/
Many thanks to the NanoTail developer for help and advice!}
}
\value{
a [tibble] containing read_classes data for
all specified samples, with metadata provided in samples_table
stored as separate columns
}
\description{
This function can be used to load any number of files with read_classes
predictions with single invocation, allowing for metadata specification.
}
\examples{
\dontrun{

samples_table <- data.frame(class_path = c(path_1,path_2,path_3,path_4),
                            sample_name =c("wt_1","mut_1","wt_2","mut_2"),
                            group = c("wt","mut","wt","mut"))

classes_data <- ninetails::read_class_multiple(samples_table)
}
}
