% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{read_class_single}
\alias{read_class_single}
\title{Reads ninetails read_classes dataframe from file.}
\usage{
read_class_single(class_path, sample_name = NA)
}
\arguments{
\item{class_path}{a character string. Path to ninetails output file}

\item{sample_name}{a sample identifier (optional), provided as a character
string. If specified will be included as an additional column sample_name.

Function based on PK (smaegol) NanoTail read_polya_single
https://github.com/LRB-IIMCB/nanotail/
Many thanks to the NanoTail developer for help and advice!}
}
\value{
a [tibble] with read_classes predictions
}
\description{
This is the basic function used to import read_classes output from check_tails
to R.
}
\examples{
\dontrun{

class_path <- "/directory/with/ninetails/read_class_output.tsv"
class_data <- ninetails::read_class_single(class_path)
}
}
