% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{read_residue_single}
\alias{read_residue_single}
\title{Reads ninetails nonadenosine_residues data from file}
\usage{
read_residue_single(residue_path, sample_name = NA)
}
\arguments{
\item{residue_path}{path to ninetails nonadenosine_residues output file}

\item{sample_name}{sample name (optional), provided as a string.
If specified will be included as an additional column sample_name.

Function based on PK (smaegol) NanoTail read_polya_single
https://github.com/LRB-IIMCB/nanotail/
Many thanks to the NanoTail developer for help and advice!}
}
\value{
a [tibble] with non-A predictions
}
\description{
This is the basic function used to import output from check_tails to R:
it imports the dataframe containing non-A detailed position info
(nonadenosine_residues).
}
\examples{
\dontrun{
residue_path <- "/directory/with/ninetails/nonadenosine_residues_output.tsv"
residue_data <- ninetails::read_residue_single(residue_path)
}
}
