% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails_guppy.R
\name{save_outputs}
\alias{save_outputs}
\title{Save pipeline outputs to files.}
\usage{
save_outputs(outputs, save_dir, prefix = "")
}
\arguments{
\item{outputs}{Named list containing data frames to save. Expected
names are \code{read_classes} and \code{nonadenosine_residues}.}

\item{save_dir}{Character string. Full path of the directory where
output files should be stored.}

\item{prefix}{Character string (optional, default \code{""}). If
provided, inserted into file names between the timestamp and the
file-type suffix.}
}
\value{
Invisible \code{NULL}. Called for the side effect of writing
  files.
}
\description{
Writes the ninetails pipeline output data frames (\code{read_classes}
and \code{nonadenosine_residues}) to tab-separated files in
\code{save_dir}. File names include a timestamp and an optional
user-defined prefix.
}
\details{
This function is called internally by
\code{\link{check_tails_guppy}} at the end of the pipeline. It is
not intended to be called directly by the user.
}
\seealso{
\code{\link{check_tails_guppy}} which calls this function.
}
\keyword{internal}
