% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_check_tails.R
\name{split_polya_data}
\alias{split_polya_data}
\title{Split large poly(A) data file into smaller polya_parts}
\usage{
split_polya_data(polya_data, part_size = 1e+05, save_dir)
}
\arguments{
\item{polya_data}{character string. Full path of the .tsv file produced
by either nanopolish polya function or tailfindr. Only DRS-derived tables are
accepted. In case of tailfindr output, it would be immediately converted to
nanopolish-like format.}

\item{part_size}{numeric [100000] (optional). If provided, defines maximum
number of rows in the poly(A) length containing input processed at once.
This value must be >=1000 (defaults to 1000000).If the size of
nanopolish/tailfindr output exceeds default 100000 rows, the input dataframe
would be splitted into parts and processed sequentially to avoid memory
overflow and pipeline crush.}

\item{save_dir}{character string. Full path of the directory where the output
files containing the tail composition information should be stored.}
}
\value{
List of file paths to created polya_parts
}
\description{
This function is not intended to be used as a standalone function, i.e., outside
the \code{\link{check_tails_guppy}} function. It does not affect data
processing; it only provides code to store results in files. It is enclosed
as a separate function, called by the wrapper, to keep the code clean and
modular, facilitating further development. Therefore, it should not be called
manually by the user.
}
\keyword{internal}
