% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_core_functions.R
\name{split_tail_centered}
\alias{split_tail_centered}
\title{Extracts fragments of poly(A) tail signal containing potential modifications
along with its delimitation (positional indices; coordinates) within the tail.}
\usage{
split_tail_centered(readname, tail_feature_list)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}
}
\value{
a list object (nested) containing 3 categories: resulting fragments
(chunk_sequence), start coordinate of given fragment (chunk_start_pos)
its' end coordinate (chunk_end_pos) arranged by the signal ID and positional
indices (WARNING! from 3' end!).
}
\description{
This function finds areas in the poly(A) tail signal containing potential
non-A residues. Extracts strings spanning 100 data points (extrapolated)
where the potential modification is always at the center of a given
extracted fragment.
}
\details{
If the data indicating the presence of modifications are in the regions
of the ends of the signal (3' or 5'), then the missing upstream or downstream
data are imputed based on the most frequent values in the entire signal.

The function returns a list object (nested) which consists of the individual
fragments of the input signal (chunk_sequence) and the coordinates
of the given fragment in the input signal: start position (chunk_start_pos)
and end position (chunk_end_pos).
}
\examples{
\dontrun{

split_tail_centered(readname= "1234-anexample-r3adn4m3",
                    tail_feature_list = tail_feature_list)

}
}
