% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ninetails_training_dataset_production_functions_guppy.R
\name{split_tail_centered_trainingset}
\alias{split_tail_centered_trainingset}
\title{Extracts fragments of poly(A) tail signal containing potential modifications
along with its delimitation (positional indices; coordinates) within the tail.}
\usage{
split_tail_centered_trainingset(readname, tail_feature_list)
}
\arguments{
\item{readname}{character string. Name of the given read within the
analyzed dataset.}

\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}
}
\value{
a list object (nested) containing 4 categories: resulting fragments
(chunk_sequence), start coordinate of given fragment (chunk_start_pos)
its' end coordinate (chunk_end_pos) and "pseudomoves" (pseudomoves)
corresponding to the given signal chunk arranged by the signal ID and
positional indices (WARNING! from 3' end!).
}
\description{
This version of the function contains an additional category (pseudomoves)
within the each chunk sublist, therefore is intended for preparation the
training/validation data.
}
\examples{
\dontrun{

split_tail_centered_trainingset(readname= "1234-anexample-r3adn4m3",
                                tail_feature_list = tail_feature_list)

}
}
