% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_with_overlaps.R
\name{split_with_overlaps}
\alias{split_with_overlaps}
\title{Splits signal to overlapping fragments of equal length. In case if the signal
is not completely divisible by given segment length, the function fills
missing data (NAs) with 3 most frequent values from the entire signal
(randomly sampled).}
\usage{
split_with_overlaps(signal, segment, overlap)
}
\arguments{
\item{signal}{character string. Name of the given ONT signal.}

\item{segment}{numeric [1]. Length of the chunk(s) to be created.}

\item{overlap}{numeric [1]. Length of the overlap between the chunks.}
}
\value{
a list object containing splitted chunks of signal.
}
\description{
Splits signal to overlapping fragments of equal length. In case if the signal
is not completely divisible by given segment length, the function fills
missing data (NAs) with 3 most frequent values from the entire signal
(randomly sampled).
}
\examples{
\dontrun{

split_with_overlaps <- function(signal='12fdcb3-ewfd543-34552-1234ddta345',
                                segment = 300, overlap 100)

}


# based on https://stackoverflow.com/questions/8872376/split-vector-with-overlapping-samples-in-r

}
