% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ninetails_training_dataset_production_functions.R
\name{split_with_overlaps}
\alias{split_with_overlaps}
\title{Splits signal to overlapping fragments of equal length.}
\usage{
split_with_overlaps(readname, tail_feature_list, segment, overlap)
}
\arguments{
\item{readname}{character string. Name of the given ONT signal.}

\item{tail_feature_list}{list object produced by create_tail_feature_list
function.}

\item{segment}{numeric [1]. Length of the chunk(s) to be created.}

\item{overlap}{numeric [1]. Length of the overlap between the chunks.}
}
\value{
a list object containing split chunks of signal.
}
\description{
In case if the signal is not completely divisible by given segment length,
the function fills missing data (NAs) with 10 most frequent values from the
entire signal (randomly sampled).
}
\examples{
\dontrun{

split_with_overlaps <- function(signal='12fdcb3-ewfd543-34552-1234ddta345',
                                segment = 300, overlap 100)

}
}
