% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_with_overlaps_moved.R
\name{split_with_overlaps_moved}
\alias{split_with_overlaps_moved}
\title{Creates vector of overlapping tail fragments extracted from
nanopolish output and fast5 files, in which move value = 1 occurs.}
\usage{
split_with_overlaps_moved(moves, signal, segment, overlap)
}
\arguments{
\item{moves}{either a numeric vector of moves or character name 
of such vector.}

\item{signal}{either a corresponding numeric vector of signal values 
or a character name of such vector.}

\item{segment}{numeric [1]. Length of the chunk(s) to be created.}

\item{overlap}{numeric [1]. Length of the overlap between the chunks.}
}
\value{
A vector of indices of moved chunks organized by their order in 
an original list is returned.
}
\description{
Creates vector of overlapping tail fragments extracted from
nanopolish output and fast5 files, in which move value = 1 occurs.
}
\examples{
\dontrun{

split_with_overlaps_moved(moves = vector_of_moves,
                          signal = vector_of_signal_corresponding_to_moves,
                          segment = 100
                          overlap = 10)

}


}
