% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{spread_nonA_residues}
\alias{spread_nonA_residues}
\title{Spreads counts of the nonA residues in modified reads in
*_nonadenosine_residues dataframe (residue data) produced by ninetails pipeline
to 3 separate columns, each containing respective C, G, U prediction
per given read.}
\usage{
spread_nonA_residues(residue_data)
}
\arguments{
\item{residue_data}{A dataframe or tibble containig non-A residue predictions
made by ninetails pipeline (filename ends with "_nonadenosine_residues.tsv")}
}
\value{
a tibble with spread prediction column (3 cols instead)
}
\description{
In addition, an extra "nonA_residues" column is located at the end of
the output table. It contains all non-A residues positions summarized (per read),
given from the 5' to 3' end, separated by ":".
}
\examples{
\dontrun{

spread_table <- ninetails::spread_nonA_residues(residue_data=residue_data)

}
}
