% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_data_postprocessing_functions.R
\name{summarize_nonA}
\alias{summarize_nonA}
\title{Produces summary table of nonA occurrences within analyzed dataset.}
\usage{
summarize_nonA(
  merged_nonA_tables,
  summary_factors = c("group"),
  transcript_id_column = c("ensembl_transcript_id_short")
)
}
\arguments{
\item{merged_nonA_tables}{an output of \code{\link{merge_nonA_tables}} function}

\item{summary_factors}{character string or vector of strings;
column(s) used for grouping (default: "group")}

\item{transcript_id_column}{character string; column with transcript id data
(default: "ensembl_transcript_id_short" as added during data preprocessing;
can be changed by the user)}
}
\value{
a summary table of nonA occurrences (tibble)
}
\description{
Creates a table with statistics for non-A residues occurrences
for each transcript (ensembl_transcrit_id_short) per each analyzed sample in given dataset.
}
\details{
In the table, "counts" are understood as the number of reads in total
or containing a given type of non-A residue (see column headers for details).
Whereas "hits" are understood as the number of occurrences of a given
nonA in total (see column headers for details). Please be aware that
there may be several "hits" in one read.

The function also reports the mean and median poly(A) tail length
by transcript.
}
\examples{
\dontrun{

summarized <- ninetails::summarize_nonA(merged_nonA_tables=merged_nonA_tables,
                                        summary_factors="group",
                                        transcript_id_column="ensembl_transcript_id_short")

}
}
