% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetails_misc_helper_functions.R
\name{winsorize_signal}
\alias{winsorize_signal}
\title{Winsorizes nanopore signal.}
\usage{
winsorize_signal(signal)
}
\arguments{
\item{signal}{numeric vector. A vector corresponding to given ONT signal.}
}
\value{
winsorized signal (numeric vector).
}
\description{
Removes high cliffs (extending above & below
signal range). Slightly affects signal extremes.
}
\details{
Based on the A. Signorell's DescTools
https://github.com/AndriSignorell/DescTools/blob/master/R/DescTools.r
}
\examples{
\dontrun{

winsorized_signal <- winsorize_signal(signal= sample(200:300))

}

}
